Copyright added, declaration of hashmap changed, some minor changes on logic of removing.
