+++ substratevm/src/com.oracle.svm.hosted/src/com/oracle/svm/hosted/phases/RemovingUnneededClassInitNodesPhase.java
@@ -1,0 +1,52 @@
+/*
+ * Copyright (c) 2011, 2020, Oracle and/or its affiliates. All rights reserved.
+ * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
+ *
+ * This code is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License version 2 only, as
+ * published by the Free Software Foundation.  Oracle designates this
+ * particular file as subject to the "Classpath" exception as provided
+ * by Oracle in the LICENSE file that accompanied this code.
+ *
+ * This code is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
+ * version 2 for more details (a copy is included in the LICENSE file that
+ * accompanied this code).
+ *
+ * You should have received a copy of the GNU General Public License version
+ * 2 along with this work; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
+ *
+ * Please contact Oracle, 500 Oracle Parkway, Redwood Shores, CA 94065 USA
+ * or visit www.oracle.com if you need additional information or have any
+ * questions.
+ */
+package com.oracle.svm.hosted.phases;
+
+import org.graalvm.compiler.nodes.AbstractBeginNode;
+import org.graalvm.compiler.nodes.InvokeWithExceptionNode;
+import org.graalvm.compiler.nodes.StructuredGraph;
+import org.graalvm.compiler.nodes.cfg.ControlFlowGraph;
+import org.graalvm.compiler.phases.Phase;
+
+import java.util.Map;
+import java.util.Set;
+
+public class RemovingUnneededClassInitNodesPhase extends Phase {
+
+    @Override
+    protected void run(StructuredGraph graph) {
+        RemovingUnneededClassInitClosure removingUnneededClassInitClosure = new RemovingUnneededClassInitClosure();
+        ControlFlowGraph cfg = ControlFlowGraph.compute(graph, true, true, true, false);
+        removingUnneededClassInitClosure.removeUnneededNodes(removingUnneededClassInitClosure, cfg.getStartBlock());
+
+        Set<Map.Entry<InvokeWithExceptionNode, AbstractBeginNode>> nodesToBeRemoved = removingUnneededClassInitClosure.getNodesToBeRemoved().entrySet();
+        for (Map.Entry<InvokeWithExceptionNode, AbstractBeginNode> entry : nodesToBeRemoved) {
+            InvokeWithExceptionNode invokeWithExceptionNode = entry.getKey();
+            AbstractBeginNode survivingSuccessor = invokeWithExceptionNode.killKillingBegin();
+
+            invokeWithExceptionNode.graph().removeSplitPropagate(invokeWithExceptionNode, survivingSuccessor);
+        }
+    }
+}
